@@ -125,7 +125,7 @@
     "pg": "^8.7.3",
     "prom-client": "^14.0.1",
     "querystring": "^0.2.1",
-    "redis": "^4.1.0",
+    "redis": "^4.1.1",
     "superagent": "^7.1.6",
     "uuid": "^8.3.2"
   },
@@ -142,7 +142,7 @@
     "@types/http-errors": "^1.8.2",
     "@types/jest": "^27.5.2",
     "@types/jsonwebtoken": "^8.5.8",
-    "@types/node": "^16.11.41",
+    "@types/node": "^16.11.43",
     "@types/nunjucks": "^3.2.1",
     "@types/passport": "^1.0.9",
     "@types/passport-oauth2": "^1.4.11",
@@ -150,20 +150,20 @@
     "@types/superagent": "^4.1.15",
     "@types/supertest": "^2.0.12",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.30.0",
-    "@typescript-eslint/parser": "^5.30.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.4",
+    "@typescript-eslint/parser": "^5.30.4",
     "concurrently": "^7.2.2",
     "cookie-session": "^2.0.0",
     "cypress": "^10.3.0",
     "cypress-multi-reporters": "^1.6.1",
-    "eslint": "^8.18.0",
+    "eslint": "^8.19.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^2.7.1",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-no-only-tests": "^2.6.0",
-    "eslint-plugin-prettier": "^4.1.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.1",
     "jest": "^27.5.1",
     "jest-html-reporter": "^3.5.0",
@@ -171,11 +171,11 @@
     "lint-staged": "^13.0.3",
     "mocha": "^10.0.0",
     "mocha-junit-reporter": "^2.0.2",
-    "nock": "^13.2.7",
+    "nock": "^13.2.8",
     "nodemon": "^2.0.18",
     "prettier": "^2.7.1",
     "sass": "^1.53.0",
-    "supertest": "^6.2.3",
+    "supertest": "^6.2.4",
     "ts-jest": "^27.1.5",
     "typescript": "^4.7.4"
   }
